+++ lib/src/db/isoexpress/rt_energy_offer.dart
@@ -173,13 +173,11 @@ class RtEnergyOfferArchive {
   }
 
   ///
-  int updateDuckDb([List<Month>? months]) {
-    final home = Platform.environment['HOME'];
-    final con =
-        Connection('$home/Downloads/Archive/IsoExpress/energy_offers.duckdb');
+  int updateDuckDb({required List<Month> months, required String pathDbFile}) {
+    final con = Connection(pathDbFile);
     con.execute('''
-CREATE TABLE IF NOT EXISTS rt_energy_offers (
-    HourBeginning TIMESTAMP_S NOT NULL,
+CREATE TABLE IF NOT EXISTS rt_offers (
+    HourBeginning TIMESTAMPTZ NOT NULL,
     MaskedParticipantId UINTEGER NOT NULL,
     MaskedAssetId UINTEGER NOT NULL,
     MustTakeEnergy FLOAT NOT NULL,
@@ -198,16 +196,22 @@ CREATE TABLE IF NOT EXISTS rt_energy_offers (
     UnitStatus ENUM('ECONOMIC', 'UNAVAILABLE', 'MUST_RUN') NOT NULL,
 );  
   ''');
-    if (months != null) {
-      /// TODO!
-    } else {
+      for (var month in months) {
+      // remove the data if it's already there
       con.execute('''
-INSERT INTO rt_energy_offers
+DELETE FROM rt_offers 
+WHERE HourBeginning >= '${month.startDate}'
+AND HourBeginning < '${month.next.startDate}';
+      ''');
+      // reinsert the data
+      con.execute('''
+INSERT INTO rt_offers
 FROM read_csv(
-    '$dir/month/rt_energy_offers_*.csv.gz', 
+    '$dir/month/rt_energy_offers_${month.toIso8601String()}.csv.gz', 
     header = true, 
     timestampformat = '%Y-%m-%dT%H:%M:%S.000%z');
 ''');
+      log.info('   Inserted month ${month.toIso8601String()} into DuckDb');
     }
     con.close();