+++ CHANGELOG.md
@@ -5,9 +5,11 @@
 - Zip the json files for lower disk usage.  For daily archives, keep one 
   zip file for a year worth of data.  
 
+## Release 2024-08-03
+- Cleanup ISONE DA/RT energy offers DuckDB process
+
 ## Release 2024-08-02
 - Update dependencies
-
 
 ## Release 2024-07-29
 - Work on ISONE MRA capacity bid offers and results archive.  Put them